requirements.txt
streamlit
numpy
torch 
torchvision 
pillow
tensorflow
tensorflow-hub
huggingface_hub==0.15.1
git+https://github.com/sberbank-ai/Real-ESRGAN.git